tests for improved to_parquet
