+++ buckaroo/serialization_utils.py
@@ -1,8 +1,12 @@
+from io import BytesIO
 import json
 import pandas as pd
 from typing import Union, Any
 from pandas._libs.tslibs import timezones
 from pandas.core.dtypes.dtypes import DatetimeTZDtype
+from fastparquet import json as fp_json
+import logging
+logger = logging.getLogger()
 
 def is_col_dt_safe(col_or_index):
     if isinstance(col_or_index.dtype, DatetimeTZDtype):
@@ -145,3 +149,53 @@ def pd_to_obj(df:pd.DataFrame):
     else:
         obj = json.loads(df.to_json(orient='table', indent=2, default_handler=str))
     return obj['data']
+
+
+class MyJsonImpl(fp_json.BaseImpl):
+    def __init__(self):
+        import json
+
+        #logger.debug("Using json encoder/decoder")
+        self.api = json
+
+    def dumps(self, data):
+        from pandas._libs.json import ujson_dumps
+        return ujson_dumps(data, default_handler=str).encode("utf-8")
+
+    def loads(self, s):
+        return self.api.loads(s)
+
+
+def to_parquet(df):
+    data: BytesIO = BytesIO()
+    
+    orig_close = data.close
+    data.close = lambda: None
+    # I don't like this copy.  modify to keep the same data with different names
+    df2 = df.copy()
+    df2['index'] = df2.index
+    df2.columns = [str(x) for x in df2.columns]
+    obj_columns = df2.select_dtypes([pd.CategoricalDtype(), 'object']).columns.to_list()
+    encodings = {k:'json' for k in obj_columns}
+
+    orig_codec_classes = fp_json._codec_classes
+    orig_get_cached_codec = fp_json._get_cached_codec
+    #fp_json._codec_classes = {'json':MyJsonImpl}
+    def fake_get_cached_codec():
+        return MyJsonImpl()
+
+    fp_json._get_cached_codec = fake_get_cached_codec
+    try:
+        df2.to_parquet(data, engine='fastparquet', object_encoding=encodings)
+    except Exception as e:
+        logger.error("error serializing to parquet %r", e)
+        raise
+    finally:
+        data.close = orig_close
+    fp_json._codec_classes = orig_codec_classes
+    fp_json._get_cached_codec = orig_get_cached_codec
+
+
+    data.seek(0)
+    return data.read()
+
